+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## 1.24.3 October 15, 2025
+– Updated filtering results to include "available rebates" taxonomy + added murb/god meta check.
+
 ## 1.24.2 October 15, 2025
 – Results cards styling updates with accessible labeling of icons.
 – Results messaging content and styling.